Time: 25 ms (32.40%), Space: 10.1 MB (51.20%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.cpp
@@ -3,7 +3,7 @@ public:
     vector<int> singleNumber(vector<int>& nums) {
         unsigned long long x_xor_y = 0;
         for(auto n: nums) x_xor_y ^=n;
-        unsigned long long right_most_set_bit = x_xor_y & -x_xor_y;
+        int right_most_set_bit = x_xor_y & -x_xor_y;
         int x = 0;
         int y = 0;
         for(auto n: nums){
